Split chat triggers from control
